[rdar://116401230] refactor: move components inside Hero folder
