+++ src/components/DocumentationTopic/Hero/HierarchyCollapsedItems.vue
